+++ crawl-ref/source/dgn-shoals.h
@@ -3,7 +3,7 @@
 
 void dgn_build_shoals_level(int level_number);
 void shoals_postprocess_level();
-void shoals_apply_tides(long turns_elapsed, bool force,
+void shoals_apply_tides(int turns_elapsed, bool force,
                         bool incremental_tide);
 void shoals_release_tide(monster* caller);
